last-ones

+++ util.h
@@ -7,15 +7,15 @@
 using namespace std;
 
 typedef struct Point{
-    float x, y;
+    double x, y;
 }Point;
 
 void read_file(char *filename, vector <Point> &points);
 
-float kmeans_cpu(Point* points, Point* means,int* labels,float* dist, int iter, int n, int k);
-float kmeans_gpu(Point* points, Point* cpupoints, Point* means,int* labels,float* dist,float* cpudist, int iter, int n, int k);
-float kmeans_cpu_ineq(Point* points,Point* means, int* labels,float* icd,int* rid,int iter,int n,int k);
-float kmeans_gpu_ineq(Point* points, Point* means, int* labels, float* icd, int* rid, int iter, int n, int k);
+double kmeans_cpu(Point* points, Point* means,int* labels,double* dist, int iter, int n, int k);
+double kmeans_gpu(Point* points, Point* cpupoints, Point* means,int* labels,double* dist,double* cpudist, int iter, int n, int k);
+double kmeans_cpu_ineq(Point* points,Point* means, int* labels,double* icd,int* rid,int iter,int n,int k);
+double kmeans_gpu_ineq(Point* points, Point* means, int* labels, double* icd, int* rid, int iter, int n, int k);
 
 vector <Point> kmeans_gpu(vector <Point> points, int iter);
